create new models subrepo which would have future investment
models within it
