+++ platforms/README.md
@@ -1,6 +1,6 @@
 # Platforms
 
-This sub repo contains the various platforms part of the Open Financing platform.
+This sub repo contains the various platforms part of the openx platform.
 
-1. Bonds
+1. OZones (Opportunity Zones)
 2. OpenSolar
