+++ platforms/opensolar/developer.go
@@ -1,4 +1,4 @@
-package solar
+package opensolar
 
 // this should contain the future developer related functions once we define them concretely
 func NewDeveloper(uname string, pwd string, seedpwd string, Name string,
